fix practice problems(02, 03)
